refacto: split @pixano/core/api into multiple files
