+++ ui/components/core/src/api/index.ts
@@ -1,0 +1,11 @@
+export * from "./addSchema";
+export * from "./deleteSchemasByIds";
+export * from "./getBrowser";
+export * from "./getDataset";
+export * from "./getDatasetItem";
+export * from "./getDatasetItemsIds";
+export * from "./getDatasetsInfo";
+export * from "./getModels";
+export * from "./getSources";
+export * from "./getViewEmbeddings";
+export * from "./updateSchema";
